rm common
