+++ benchmark/src/serial/uts.cpp
@@ -1,6 +1,5 @@
 #include <benchmark/benchmark.h>
 
-#include "common.hpp"
 #include "macros.hpp"
 #include "uts.hpp"
